Do not send random_sample as 0. Closes #290
This also changes the way of keeping corpus configuration so past corpus
configuration will be lost. It seems at some point Django started to
store bools in forms as "on" and (possibly) "off". So, suddenly, "0" was
interpreted as "on" in random_sample setting, breaking the queries.

Also, as Poliqarp reports random sample in the same way as standard one
("23 items found of 63"), Djview4poliqarp was fooled into thinking it
can fetch more items. Now it relies upon seeing "Next page" link.